+++ src/handler/question-sync.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/cndjp/qicoo-api/src/sql"
+	"github.com/cndjp/qicoo-api/src/mysqlib"
 	"github.com/go-gorp/gorp"
 	"github.com/sirupsen/logrus"
 )
@@ -24,13 +24,13 @@ func (rc *RedisClient) syncQuestion(eventID string) {
 
 	// DBからデータを取得
 	var m *gorp.DbMap
-	db, err := sql.InitMySQL()
+	db, err := mysqlib.InitMySQL()
 	if err != nil {
 		logrus.Error(err)
 		return
 	}
 
-	m = sql.MappingDBandTable(db)
+	m = mysqlib.MappingDBandTable(db)
 
 	m.AddTableWithName(Question{}, "questions")
 	defer m.Db.Close()
