+++ src/mysqlib/mysqlib.go
@@ -1,13 +1,13 @@
-package sql
+package mysqlib
 
 import (
-	goSQL "database/sql"
+	"database/sql"
 	"os"
 
 	"github.com/go-gorp/gorp"
 )
 
-func InitMySQL() (db *goSQL.DB, err error) {
+func InitMySQL() (db *sql.DB, err error) {
 	dbms := "mysql"
 	user := os.Getenv("DB_USER")
 	password := os.Getenv("DB_PASSWORD")
@@ -16,13 +16,13 @@ func InitMySQL() (db *goSQL.DB, err error) {
 	option := "?parseTime=true"
 
 	connect := user + ":" + password + "@" + protocol + "/" + dbname + option
-	db, err = goSQL.Open(dbms, connect)
+	db, err = sql.Open(dbms, connect)
 
 	return
 }
 
 // InitMySQLDB 環境変数を利用しDBへのConnectionを取得する(sqldriverでconnection poolが実装されているらしい)
-func MappingDBandTable(db *goSQL.DB) *gorp.DbMap {
+func MappingDBandTable(db *sql.DB) *gorp.DbMap {
 	// structの構造体とDBのTableを紐づける
 	return &gorp.DbMap{Db: db, Dialect: gorp.MySQLDialect{}}
 }
